Consolidate enviroment-diagnostics into solution-problems
