+++ src/solutions/solution-problems.ts
@@ -19,11 +19,23 @@ import * as vscode from 'vscode';
 import { constructor } from '../generic/constructor';
 import { LogMessages } from '../json-rpc/csolution-rpc-client';
 import { Severity } from './constants';
+import * as fsUtils from '../utils/fs-utils';
 import { getFileNameFromPath } from '../utils/path-utils';
 import { stripTwoExtensions } from '../utils/string-utils';
+import { getWorkspaceFolder } from '../utils/vscode-utils';
 import { ProblemDiagnosticActionResolver } from './problem-diagnostic-action-resolver';
 import { SolutionLoadStateChangeEvent, SolutionManager } from './solution-manager';
 import { ConvertResultData, CbuildResultData, SolutionEventHub } from './solution-event-hub';
+
+interface SettingsLocation {
+    filePath: string;
+    range: vscode.Range;
+}
+
+interface EnvironmentMessage {
+    message: string;
+    severity: vscode.DiagnosticSeverity;
+}
 
 export const toolsPrefixPatterns = {
     error: /^.*error (?:cbuild|cbuild2cmake|csolution|cpackget):\s*/,
@@ -115,7 +127,9 @@ export interface SolutionProblems {
 export class SolutionProblemsImpl implements SolutionProblems {
 
     private readonly diagnosticCollection: vscode.DiagnosticCollection = vscode.languages.createDiagnosticCollection('csolution');
+    private readonly environmentDiagnosticCollection: vscode.DiagnosticCollection = vscode.languages.createDiagnosticCollection('csolution-environment');
     private readonly diagnosticActionResolver = new ProblemDiagnosticActionResolver();
+    private readonly environmentVariablesSetting = '"cmsis-csolution.environmentVariables"';
     /**
     *  source files for diagnostics mapping
     */
@@ -133,6 +147,7 @@ export class SolutionProblemsImpl implements SolutionProblems {
             this.eventHub.onDidCbuildCompleted(this.handleCbuildCompleted, this),
             this.solutionManager.onDidChangeLoadState(this.handleLoadStateChanged, this),
             this.diagnosticCollection,
+            this.environmentDiagnosticCollection,
         );
     }
 
@@ -140,14 +155,40 @@ export class SolutionProblemsImpl implements SolutionProblems {
         // Intentionally clear only on convert: convert is the canonical refresh point.
         // cbuild follows convert and should add diagnostics without wiping convert findings.
         this.clearDiagnostics();
-        await this.enrichAndUpdateDiagnostics(data.logMessages, data.toolsOutputMessages);
+        await Promise.all([
+            this.enrichAndUpdateDiagnostics(data.logMessages, data.toolsOutputMessages),
+            this.updateEnvironmentDiagnosticsFromConvert(data),
+        ]);
     }
 
     private async handleCbuildCompleted(data: CbuildResultData): Promise<void> {
         // Do not clear diagnostics here. cbuild diagnostics are additive after convert.
         // This preserves existing convert diagnostics and avoids churn from redundant clears.
         const logMessages: LogMessages = { success: true, errors: [], warnings: [], info: [] };
-        await this.enrichAndUpdateDiagnostics(logMessages, data.toolsOutputMessages);
+        await Promise.all([
+            this.enrichAndUpdateDiagnostics(logMessages, data.toolsOutputMessages),
+            this.updateEnvironmentDiagnosticsFromCbuild(data),
+        ]);
+    }
+
+    private async updateEnvironmentDiagnosticsFromConvert(data: ConvertResultData): Promise<void> {
+        const messages: EnvironmentMessage[] = [
+            ...(data.logMessages.errors ?? []).map(message => ({
+                message,
+                severity: vscode.DiagnosticSeverity.Error,
+            })),
+            ...(data.logMessages.warnings ?? []).map(message => ({
+                message,
+                severity: vscode.DiagnosticSeverity.Warning,
+            })),
+            ...this.extractEnvironmentMessagesFromToolOutput(data.toolsOutputMessages),
+        ];
+        await this.updateEnvironmentDiagnostics(messages);
+    }
+
+    private async updateEnvironmentDiagnosticsFromCbuild(data: CbuildResultData): Promise<void> {
+        const messages = this.extractEnvironmentMessagesFromToolOutput(data.toolsOutputMessages);
+        await this.updateEnvironmentDiagnostics(messages);
     }
 
     private async enrichAndUpdateDiagnostics(logMessages: LogMessages, toolsOutputMessages?: string[]): Promise<void> {
@@ -159,6 +200,111 @@ export class SolutionProblemsImpl implements SolutionProblems {
         if (data.previousState.solutionPath !== data.newState.solutionPath) {
             this.clearDiagnostics();
         }
+    }
+
+    private extractEnvironmentMessagesFromToolOutput(lines?: string[]): EnvironmentMessage[] {
+        if (!lines) {
+            return [];
+        }
+
+        return lines
+            .flatMap(line => {
+                if (toolsPrefixPatterns.error.test(line)) {
+                    return [{
+                        message: line.replace(toolsPrefixPatterns.error, '').trim(),
+                        severity: vscode.DiagnosticSeverity.Error,
+                    }];
+                }
+                if (toolsPrefixPatterns.warning.test(line)) {
+                    return [{
+                        message: line.replace(toolsPrefixPatterns.warning, '').trim(),
+                        severity: vscode.DiagnosticSeverity.Warning,
+                    }];
+                }
+                return [];
+            })
+            .filter(item => isEnvironmentVariableMessage(item.message));
+    }
+
+    private async updateEnvironmentDiagnostics(rawMessages: EnvironmentMessage[]): Promise<void> {
+        this.environmentDiagnosticCollection.clear();
+
+        const messages = new Map<string, vscode.DiagnosticSeverity>();
+        for (const rawMessage of rawMessages) {
+            const message = normalizeEnvironmentMessage(rawMessage.message);
+            if (!isEnvironmentVariableMessage(message)) {
+                continue;
+            }
+            const currentSeverity = messages.get(message);
+            if (currentSeverity === undefined || rawMessage.severity < currentSeverity) {
+                messages.set(message, rawMessage.severity);
+            }
+        }
+
+        if (messages.size === 0) {
+            return;
+        }
+
+        const settings = await this.getSettingsLocation();
+        if (!settings) {
+            return;
+        }
+
+        const diagnostics: vscode.Diagnostic[] = [];
+        for (const [message, severity] of messages) {
+            const fullMessage = `${message}; review ${this.environmentVariablesSetting}`;
+            const action = this.diagnosticActionResolver.resolve({
+                message: fullMessage,
+                diagnosticFilePath: settings.filePath,
+                hasLocation: true,
+            });
+
+            const entry = new vscode.Diagnostic(settings.range, action?.message ?? fullMessage, severity);
+            entry.source = 'csolution';
+            if (action?.code) {
+                entry.code = action.code;
+            }
+            diagnostics.push(entry);
+        }
+
+        const uri = vscode.Uri.file(settings.filePath);
+        this.environmentDiagnosticCollection.set(uri, diagnostics);
+        await vscode.commands.executeCommand('workbench.actions.view.problems', { preserveFocus: true });
+    }
+
+    private async getSettingsLocation(): Promise<SettingsLocation | undefined> {
+        const workspaceFolder = getWorkspaceFolder();
+        if (!workspaceFolder) {
+            return undefined;
+        }
+
+        const settingsFile = vscode.workspace.workspaceFile?.fsPath ?? path.join(workspaceFolder, '.vscode', 'settings.json');
+        if (!fsUtils.fileExists(settingsFile)) {
+            return {
+                filePath: settingsFile,
+                range: new vscode.Range(0, 0, 0, 0),
+            };
+        }
+
+        try {
+            const doc = await vscode.workspace.openTextDocument(settingsFile);
+            const startOffset = doc.getText().indexOf(this.environmentVariablesSetting);
+            if (startOffset >= 0) {
+                const pos = doc.positionAt(startOffset);
+                const endCharacter = doc.lineAt(pos.line).range.end.character;
+                return {
+                    filePath: settingsFile,
+                    range: new vscode.Range(pos.line, pos.character, pos.line, endCharacter),
+                };
+            }
+        } catch {
+            // Keep default range when settings document cannot be opened.
+        }
+
+        return {
+            filePath: settingsFile,
+            range: new vscode.Range(0, 0, 0, 0),
+        };
     }
 
     /**
